test UnwrapEthereumMsg
