+++ x/evm/types/msg.go
@@ -40,23 +40,6 @@ func NewTx(
 
 }
 
-// UnwrapEthereumMsg extract MsgEthereumTx from wrapping sdk.Tx
-func UnwrapEthereumMsg(tx *sdk.Tx) (*MsgEthereumTx, error) {
-	if tx == nil {
-		return nil, fmt.Errorf("invalid tx: nil")
-	}
-
-	if len((*tx).GetMsgs()) != 1 {
-		return nil, fmt.Errorf("invalid tx type: %T", tx)
-	}
-	msg, ok := (*tx).GetMsgs()[0].(*MsgEthereumTx)
-	if !ok {
-		return nil, fmt.Errorf("invalid tx type: %T", tx)
-	}
-
-	return msg, nil
-}
-
 // NewTxContract returns a reference to a new Ethereum transaction
 // message designated for contract creation.
 func NewTxContract(
