+++ x/evm/types/utils_test.go
@@ -1,11 +1,17 @@
-package types
+package types_test
 
 import (
+	"math/big"
 	"testing"
 
+	"github.com/cosmos/cosmos-sdk/client"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	proto "github.com/gogo/protobuf/proto"
+	"github.com/tharsis/ethermint/app"
 	"github.com/tharsis/ethermint/crypto/ethsecp256k1"
+	"github.com/tharsis/ethermint/encoding"
+	evmtypes "github.com/tharsis/ethermint/x/evm/types"
 
 	"github.com/stretchr/testify/require"
 
@@ -27,9 +33,9 @@ func GenerateEthAddress() ethcmn.Address {
 func TestEvmDataEncoding(t *testing.T) {
 	ret := []byte{0x5, 0x8}
 
-	data := &MsgEthereumTxResponse{
+	data := &evmtypes.MsgEthereumTxResponse{
 		Hash: common.BytesToHash([]byte("hash")).String(),
-		Logs: []*Log{{
+		Logs: []*evmtypes.Log{{
 			Data:        []byte{1, 2, 3, 4},
 			BlockNumber: 17,
 		}},
@@ -40,15 +46,36 @@ func TestEvmDataEncoding(t *testing.T) {
 	require.NoError(t, err)
 
 	txData := &sdk.TxMsgData{
-		Data: []*sdk.MsgData{{MsgType: TypeMsgEthereumTx, Data: enc}},
+		Data: []*sdk.MsgData{{MsgType: evmtypes.TypeMsgEthereumTx, Data: enc}},
 	}
 
 	txDataBz, err := proto.Marshal(txData)
 	require.NoError(t, err)
 
-	res, err := DecodeTxResponse(txDataBz)
+	res, err := evmtypes.DecodeTxResponse(txDataBz)
 	require.NoError(t, err)
 	require.NotNil(t, res)
 	require.Equal(t, data.Logs, res.Logs)
 	require.Equal(t, ret, res.Ret)
+}
+
+func TestUnwrapEthererumMsg(t *testing.T) {
+	_, err := evmtypes.UnwrapEthereumMsg(nil)
+	require.NotNil(t, err)
+
+	encodingConfig := encoding.MakeConfig(app.ModuleBasics)
+	clientCtx := client.Context{}.WithTxConfig(encodingConfig.TxConfig)
+	builder, _ := clientCtx.TxConfig.NewTxBuilder().(authtx.ExtensionOptionsTxBuilder)
+
+	tx := builder.GetTx().(sdk.Tx)
+	_, err = evmtypes.UnwrapEthereumMsg(&tx)
+	require.NotNil(t, err)
+
+	msg := evmtypes.NewTx(big.NewInt(1), 0, &common.Address{}, big.NewInt(0), 0, big.NewInt(0), []byte{}, nil)
+	err = builder.SetMsgs(msg)
+
+	tx = builder.GetTx().(sdk.Tx)
+	msg_, err := evmtypes.UnwrapEthereumMsg(&tx)
+	require.Nil(t, err)
+	require.Equal(t, msg_, msg)
 }
